Move Gradle public API to ArchUnitFixture
