+++ subprojects/architecture-test/src/test/java/org/gradle/architecture/test/ArchUnitFixture.java
@@ -16,11 +16,62 @@
 
 package org.gradle.architecture.test;
 
+import com.tngtech.archunit.base.DescribedPredicate;
+import com.tngtech.archunit.base.PackageMatchers;
+import com.tngtech.archunit.core.domain.JavaClass;
 import com.tngtech.archunit.lang.ArchRule;
 import com.tngtech.archunit.library.freeze.FreezingArchRule;
+import org.gradle.util.EmptyStatement;
+import org.gradle.util.Matchers;
+import org.gradle.util.PreconditionVerifier;
+import org.gradle.util.Requires;
+import org.gradle.util.SetSystemProperties;
+import org.gradle.util.TestClassLoader;
+import org.gradle.util.TestPrecondition;
+import org.gradle.util.TestPreconditionExtension;
+import org.gradle.util.UsesNativeServices;
+import org.gradle.util.UsesNativeServicesExtension;
+
+import java.util.Arrays;
+import java.util.Set;
+
+import static com.tngtech.archunit.base.DescribedPredicate.not;
+import static com.tngtech.archunit.core.domain.JavaClass.Predicates.resideInAnyPackage;
+import static java.util.stream.Collectors.toSet;
 
 public interface ArchUnitFixture {
     static ArchRule freeze(ArchRule rule) {
         return new FreezeInstructionsPrintingArchRule(FreezingArchRule.freeze(rule));
+    }
+
+    static DescribedPredicate<JavaClass> gradlePublicApi() {
+        return new GradlePublicApi();
+    }
+
+    static DescribedPredicate<JavaClass> gradleInternalApi() {
+        return resideInAnyPackage("org.gradle..")
+            .and(not(gradlePublicApi()))
+            .as("Gradle Internal API");
+    }
+
+    class GradlePublicApi extends DescribedPredicate<JavaClass> {
+        private static final PackageMatchers INCLUDES = PackageMatchers.of(parsePackageMatcher(System.getProperty("org.gradle.public.api.includes")));
+        private static final PackageMatchers EXCLUDES = PackageMatchers.of(parsePackageMatcher(System.getProperty("org.gradle.public.api.excludes")));
+        private static final DescribedPredicate<JavaClass> TEST_FIXTURES = JavaClass.Predicates.belongToAnyOf(EmptyStatement.class, Matchers.class, PreconditionVerifier.class, Requires.class, SetSystemProperties.class, TestClassLoader.class, TestPrecondition.class, TestPreconditionExtension.class, UsesNativeServices.class, UsesNativeServicesExtension.class);
+
+        public GradlePublicApi() {
+            super("Gradle public API");
+        }
+
+        @Override
+        public boolean apply(JavaClass input) {
+            return INCLUDES.apply(input.getPackageName()) && !EXCLUDES.apply(input.getPackageName()) && !TEST_FIXTURES.apply(input);
+        }
+
+        private static Set<String> parsePackageMatcher(String packageList) {
+            return Arrays.stream(packageList.split(":"))
+                .map(include -> include.replace("**/", "..").replace("/**", "..").replace("/*", "").replace("/", "."))
+                .collect(toSet());
+        }
     }
 }
